I added yellowGoLeft and yellowGoRight :) -etan
ALSO weh added chanjed the intengern to doubel
RAAHH RAAHH RAAHH RAAHH RAAHH RAAHH RAAHH RAAHH RAAHH
GAAH GAHG OOH LA LAA LAA A MOO MOO MO PHANE
-definteltley not lagon

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/MecanumDriveChassis.java
@@ -451,7 +451,7 @@ public class MecanumDriveChassis
   }
   
   
-  public void strafeRight(int distanceCm)
+  public void strafeRight(double distanceCm)
   {
     telemetry.addLine("strafing right");
     telemetry.update();
